requirements.txt
# Core dependencies
numpy>=1.21.0
numba>=0.54.0
pandas>=1.3.0
python-binance>=1.0.0

# Configuration
pyyaml>=5.4.0
python-dotenv>=0.19.0

# Async and networking
aiohttp>=3.8.0
requests>=2.26.0
websocket-client>=1.0.0

# Caching
redis>=4.0.0

# Development (optional)
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
